Refactor and Clean Up Code
